requirements.txt
streamlit
pdfminer.six
python-docx
scikit-learn
nltk
